fix: do not export matchResponse
