+++ src/common/index.ts
@@ -3,5 +3,4 @@ export { parseBody } from "./parse-body";
 export { parseCookies } from "./parse-cookies";
 export { parseDelay } from "./parse-delay";
 export { selectResponse } from "./select-response";
-export { matchResponse } from "./match-response";
 export { appendBasePath } from "./append-base-path";
